pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imalink-core"
version = "1.0.0"
description = "Core image processing library for ImaLink photo management system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Kjell Kåre Olsen"}
]
keywords = ["image", "exif", "metadata", "photo", "imalink", "photography"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
# RAW format support (optional)
raw = [
    "rawpy>=0.18.0",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

# All extras
all = [
    "imalink-core[raw,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/kjelkols/imalink-core"
Documentation = "https://github.com/kjelkols/imalink-core/blob/main/README.md"
Repository = "https://github.com/kjelkols/imalink-core"
Issues = "https://github.com/kjelkols/imalink-core/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/imalink_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=imalink_core --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
